Update and rename forbidden_gates.js to forbidden-gates-init.js
